requirements.txt
google-cloud-bigquery==2.34.0
flask==2.1.1
google-cloud-pubsub==2.10.0
pandas==1.3.5
numpy==1.21.6
requests==2.26.0
python-dotenv==0.19.2
matplotlib==3.5.1
plotly==5.1.0